requirements.txt
iso8601
python-dateutil
elasticsearch5
elasticsearch-dsl>=5.0.0,<6.0.0
attrs==17.4.0
click==6.7
Flask==0.12.2
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pluggy==0.6.0
py==1.5.2
pyparsing==2.2.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
six==1.11.0
SPARQLWrapper==1.8.0
Werkzeug==0.14.1
futures==3.1.1
google-api-python-client==1.6.5
grpcio==1.10.0
grpcio-tools==1.10.0
requests==2.18.4
